# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CodeArena competitive programming platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List the current user's submissions
      parameters:
        - name: problemId
          in: query
          required: false
          schema:
            type: integer
        - name: contestId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit code for judging (queues async Judge0 execution)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Invalid submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a submission with per-test-case results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SubmissionStatus:
      type: string
      enum:
        - queued
        - judging
        - accepted
        - partial
        - wrong_answer
        - time_limit_exceeded
        - memory_limit_exceeded
        - runtime_error
        - compilation_error
        - internal_error
    TestResultStatus:
      type: string
      enum:
        - queued
        - processing
        - accepted
        - wrong_answer
        - time_limit_exceeded
        - memory_limit_exceeded
        - runtime_error
        - compilation_error
        - internal_error
    SubmissionTestResult:
      type: object
      required: [id, testCaseId, status, isSample, points, earnedPoints]
      properties:
        id:
          type: integer
        testCaseId:
          type: integer
        status:
          $ref: "#/components/schemas/TestResultStatus"
        isSample:
          type: boolean
        points:
          type: integer
        earnedPoints:
          type: integer
        timeMs:
          type: [number, 'null']
        memoryKb:
          type: [integer, 'null']
        stdout:
          type: [string, 'null']
        stderr:
          type: [string, 'null']
        compileOutput:
          type: [string, 'null']
    SubmissionInput:
      type: object
      required: [problemId, languageId, sourceCode]
      properties:
        problemId:
          type: integer
        contestId:
          type: [integer, 'null']
        languageId:
          type: integer
        sourceCode:
          type: string
          minLength: 1
    SubmissionSummary:
      type: object
      required: [id, problemId, problemTitle, languageId, languageName, status, score, maxScore, submittedAt]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        problemTitle:
          type: string
        contestId:
          type: [integer, 'null']
        languageId:
          type: integer
        languageName:
          type: string
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        score:
          type: integer
        maxScore:
          type: integer
        submittedAt:
          type: string
          format: date-time
    Submission:
      type: object
      required: [id, problemId, problemTitle, languageId, languageName, sourceCode, status, score, maxScore, submittedAt, testResults]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        problemTitle:
          type: string
        contestId:
          type: [integer, 'null']
        languageId:
          type: integer
        languageName:
          type: string
        sourceCode:
          type: string
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        score:
          type: integer
        maxScore:
          type: integer
        submittedAt:
          type: string
          format: date-time
        judgedAt:
          type: [string, 'null']
          format: date-time
        testResults:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionTestResult"
...