# lib/api-zod/src/generated/types/testResultStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeArena competitive programming platform API
 * OpenAPI spec version: 0.1.0
 */

export type TestResultStatus = typeof TestResultStatus[keyof typeof TestResultStatus];


export const TestResultStatus = {
  queued: 'queued',
  processing: 'processing',
  accepted: 'accepted',
  wrong_answer: 'wrong_answer',
  time_limit_exceeded: 'time_limit_exceeded',
  memory_limit_exceeded: 'memory_limit_exceeded',
  runtime_error: 'runtime_error',
  compilation_error: 'compilation_error',
  internal_error: 'internal_error',
} as const;